Updated Camera helpers and weapon_item_spawn
Deleted the sprite helper models. They caused errors soemtimes. Instead we got the default cubes back because any model is bad. So now the oriented sprite uses a "frustum" to show which way it points.

The camera helper models are updated to show a fov frustum that changes size when the fov is changed.
Axis helpers got reworked. The arrows are easier to see now.