various small edits
Changed env_fade to add a flag it was missing.
Edited doors default settings.
Changed order of all entities so the default base is "parentname, global name, targetname, angles" in that order, if applicable. This just means that the name field and parent name fiels will appear in the same order oeverywhere.