Important cubemap fix
First and foremost, the important part is removing "Origin" from all cubemaps.

Placing cubemaps inside instances while using the ORIGIN value causes them to break and be moved to the map center!
I just figured this out, which broke my map and causes immense headaches.

Furthermore, removing the value via smart edit causes all cubemaps to be moved to the map center too.

Other than that, i also added "Trigger" as a base to trigger_tonemap, as it is a trigger but didn't work like one, but could.